update to 1.19.4

+++ src/main/java/ladysnake/illuminations/mixin/BlockMixin.java
@@ -12,8 +12,6 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import java.util.Random;
-
 @Mixin(Block.class)
 public abstract class BlockMixin {
     @Shadow
@@ -23,7 +21,7 @@ public abstract class BlockMixin {
     protected void illuminations$randomDisplayTick(BlockState state, World world, BlockPos pos, net.minecraft.util.math.random.Random random, CallbackInfo ci) {
         if (this.getDefaultState().getBlock() == Blocks.SEA_LANTERN) {
             for (int i = 0; i < 10; i++) {
-                BlockPos blockPos = new BlockPos(pos.getX() + 0.5 + random.nextGaussian() * 15, pos.getY() + 0.5 + random.nextGaussian() * 15, pos.getZ() + 0.5 + random.nextGaussian() * 15);
+                BlockPos blockPos = new BlockPos((int) (pos.getX() + 0.5 + random.nextGaussian() * 15), (int) (pos.getY() + 0.5 + random.nextGaussian() * 15), (int) (pos.getZ() + 0.5 + random.nextGaussian() * 15));
 
                 if (world.getBlockState(blockPos).getBlock() == Blocks.WATER && random.nextInt(1 + world.getLightLevel(blockPos)) == 0) {
                     world.addParticle(Illuminations.PRISMARINE_CRYSTAL, true, blockPos.getX(), blockPos.getY(), blockPos.getZ(), 0f, 0f, 0f);
